+++ Changelog.md
@@ -6,6 +6,9 @@
    programmatic rendering of various shapes in a scene.
     * [Pull request 2498](https://bitbucket.org/osrf/gazebo/pull-request/2498)
 
+1. Support version 5 of the DART Physics Engine.
+    * [Pull request #2459](https://bitbucket.org/osrf/gazebo/pull-request/2459)
+
 1. UserCamera overrides `Camera::Render` to reduce CPU usage.
     * [Pull request 2480](https://bitbucket.org/osrf/gazebo/pull-request/2480)